+++ lib/server/http_simulator_api.dart
@@ -4,7 +4,7 @@ import 'dart:isolate';
 import 'package:collection/collection.dart';
 import 'package:flutter/foundation.dart';
 import 'package:knowgo/api.dart' as knowgo;
-import 'package:knowgo_vehicle_simulator/server/http_exve_api.dart';
+import 'package:knowgo_vehicle_simulator/server.dart';
 import 'package:knowgo_vehicle_simulator/simulator.dart';
 import 'package:shelf/shelf.dart';
 import 'package:shelf_router/shelf_router.dart';