+++ lib/views.dart
@@ -1,1 +1,2 @@
+export 'views/event_injection_home.dart';
 export 'views/vehicle_simulator_home.dart';